requirements.txt
discord.py
requests
flask
<<<<<<< HEAD
gunicorn
requests
=======
python-dotenv
>>>>>>> 4c40460 (test)
